refactor further
